Initial commit of the PHP-Gremlin-Native Skeleton
This library strives to be a one-to-one mapping of the core functionality of the Gremlin language in Tinkerpop to enable users to natively work in the code without needing to hand-type Gremlin queries. This will ultimately be compiled into ByteCode, and functionally useful as a library with the Gremlin-PHP server integration.